Improve organization. Start documentation
