+++ OmniAX/Sources/Core/Traits.swift
