+++ OmniAX/Sources/Extensions/String+AX.swift
