+++ OmniAX/Sources/Managed Reference/WrappedHashable.swift
